feat(nlp): add NLP processor and utility functions for text analysis
