+++ packages/utils/package.json
@@ -11,6 +11,7 @@
     "./time": "./src/time.ts",
     "./location": "./src/location/index.ts",
     "./logger": "./src/logging/logger.ts",
+    "./nlp": "./src/nlp/index.ts",
     "./google-sheets": "./src/google/sheets.ts",
     "./google-calendar": "./src/google/calendar.ts",
     "./google-maps": "./src/google/maps.ts",
