+++ packages/utils/src/index.ts
@@ -3,5 +3,5 @@ export * from './logging/logger'
 // Utils
 export * from './rate-limit'
 export * from './time'
-export * from './types'
+export * from './types/index'
 export * from './writer'
